Move `FeatureThemeProvider` to `:core:ui:theme:api` module
