+++ app-thunderbird/src/main/kotlin/net/thunderbird/android/provider/TbFeatureThemeProvider.kt
@@ -2,9 +2,9 @@ package net.thunderbird.android.provider
 
 import androidx.compose.runtime.Composable
 import app.k9mail.core.ui.compose.theme2.thunderbird.ThunderbirdTheme2
-import app.k9mail.feature.launcher.FeatureLauncherExternalContract
+import app.k9mail.core.ui.theme.api.FeatureThemeProvider
 
-class TbFeatureThemeProvider : FeatureLauncherExternalContract.FeatureThemeProvider {
+class TbFeatureThemeProvider : FeatureThemeProvider {
     @Composable
     override fun WithTheme(content: @Composable () -> Unit) {
         ThunderbirdTheme2 {
