+++ core/ui/theme/README.md
@@ -1,0 +1,3 @@
+## Theme
+
+The app theme is provided by the `ThemeProvider` found in the `api` module and should be implemented in the app modules with the required app themes from the `core:compose` and `core:legacy` modules.
